Midiendo OK

+++ app/src/main/java/com/index/medidor/utils/BluetoothDataReceiver.java
@@ -2,5 +2,5 @@ package com.index.medidor.utils;
 
 public interface BluetoothDataReceiver {
 
-    void getBluetoothData(int dato);
+    void getBluetoothData(double dato);
 }
